requirements/dev.txt
django==2.0.7
django-extensions==2.0.7
psycopg2==2.7.5
pygraphviz==1.3.1
pymysql==0.9.2
mysqlclient==1.3.13
## Comentar la linea 4 y descomentar la linea 8, instalar los requerimientos de nuevo
## --install-option="--include-path=/usr/local/include/graphviz/" --install-option="--library-path=/usr/local/lib/graphviz" pygraphviz
djangorestframework==3.9.0

# Instalar últimas versiones
#django
#psycopg2
#django-extensions
#pygraphviz
#pymysql
#mysqlclient
#djangorestframework
